+++ dimos/mapping/ray_tracing/rust/src/module.rs
@@ -249,7 +249,7 @@ fn make_cloud(data: Vec<u8>, n: u32, frame_id: &str, stamp: Time) -> PointCloud2
     let make_field = |name: &str, off: u32| PointField {
         name: name.into(),
         offset: off,
-        datatype: PointField::FLOAT32 as u8,
+        datatype: PointField::FLOAT32,
         count: 1,
     };
     PointCloud2 {
